+++ src/main/webapp/WEB-INF/views/customer/qnaBoardView.jsp
@@ -49,10 +49,24 @@
         </div>
         <div class="button">
         <input type="button" value="목록" onclick="showQnaBoardList()"/>
+        
+	<% 	if(
+				loginMember != null && 
+				(
+				  loginMember.getMemberId().equals(qnaBoard.getWriter())
+				  || MemberService.ADMIN_ROLE.equals(loginMember.getMemberRole())
+				)
+			){ %>
         <input type="button" value="수정" onclick="updateQnaBoard()"/>
         <input type="button" value="삭제" onclick="deleteQnaBoard()"/>
+         <% 	} %>
+        <% 	if(
+				loginMember != null && 
+				( MemberService.ADMIN_ROLE.equals(loginMember.getMemberRole()))
+			){ %>
         <input type="button" value="답변" onclick="replyQnaBoard()"/>
         </div>
+  <% 	} %>
  </div>
  </section>
 <!--  게시글 삭제-->
